fix naming in support classes
