+++ java/vector/src/main/java/org/apache/arrow/vector/BaseFixedWidthVector.java
