+++ java/vector/src/main/java/org/apache/arrow/vector/BaseVariableWidthVector.java
