+++ java/vector/src/main/java/org/apache/arrow/vector/UInt2Vector.java
