+++ java/vector/src/main/java/org/apache/arrow/vector/UInt4Vector.java
